ver. 1.4.0 - Release

+++ src/src/ParticlesUtils.cpp
@@ -27,7 +27,7 @@
 //-----------------------------------------------------------------------------
 float getRandomMinMax( float fMin, float fMax )
 {
-    float fRandNum = (float)rand () / RAND_MAX;
+    float fRandNum = (float)rand () / (float)RAND_MAX;
     return fMin + (fMax - fMin) * fRandNum;
 }
